requirements.txt
streamlit
pandas
pandasql
python-dotenv
groq
semantic-router[local]
chromadb
sentence_transformers
transformers
torch
